requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
PyJWT==2.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
